backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
ollama==0.1.7
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
sqlalchemy==2.0.23
alembic==1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
textstat==0.7.3
nltk==3.8.1
aiosqlite==0.19.0
databases[sqlite]==0.8.0